+++ include/entities/ZstPerformer.h
@@ -29,7 +29,7 @@ public:
 	ZST_EXPORT const char * address();
 
 	//Creatables
-	ZST_EXPORT int num_creatables();
+	ZST_EXPORT size_t num_creatables();
 
 	ZST_EXPORT virtual void write(std::stringstream & buffer) override;
 	ZST_EXPORT virtual void read(const char * buffer, size_t length, size_t & offset) override;
